src/requirements.txt
discord.py==2.0.0
python-dotenv
opencensus-ext-azure
azure-core
azure-data-tables